Add clean and qemu test targets to Makefile.
Selecting the loader.efi in OVMF manually is not super convenient,
but at least it is a reproducible way to run the compiled EFI
binary without rebooting the whole machine.

I still cannot produce a working binary on my system, but at least,
this Makefile target allows me to fail faster.

Tested on vanilla Windows 11 with WSL Ubuntu and gnu-efi 3.0.11
and qemu 6.2.0.